Create Chocolate Chip Cookie Recipe Page
Developed a new HTML page named "chocolate-chip-cookie-recipe.html" to
showcase a delectable chocolate chip cookie recipe. Structured the
layout and design to present the recipe in a clear and organized manner,
ensuring ease of readability for users. Included detailed instructions
and ingredient list for crafting the perfect batch of chocolate chip cookies.
Incorporated an image to visually enhance the page and provide a preview
of the mouthwatering final product. Added link to chocolate chip cookie
website page to index.html. Added cookie.jpg to img directory.