Create classic caprese salad recipe page
Developed a new HTML page named "classic-caprese-salad.html" to
showcase a delectable salad recipe. Structured the
layout and design to present the recipe in a clear and organized manner,
ensuring ease of readability for users. Included detailed instructions
and ingredient list for crafting the perfect salad recipe.
Incorporated an image to visually enhance the page and provide a preview
of the mouthwatering final product. Added link to caprese salad
website page to index.html. Added caprese-salad.jpg to img directory.